Remove v2 folder and update tests
Everything that was in the v2 folder has now been folded into the v1 version, and the v1 version removed where
possible.
